rename: 유저의 활동과 관련된 페이지들 디렉토리명 수정
RegistReviewPage -> MyReviewPage
RegistedPage -> MyTreePage
SavePage -> SvaedTreePage
